requirements.txt
asttokens==2.4.1
backcall==0.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
executing==2.0.1
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.28.0
ipython==8.12.3
jedi==0.19.1
jupyter-client==8.6.0
jupyter-core==5.6.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.33
nest-asyncio==1.5.8
numpy==1.24.4
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.12
zipp==3.17.0
